requirements.txt
typer>=0.9.0
rich>=13.0.0
playwright>=1.40.0
httpx>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
markdown>=3.5.0
lancedb>=0.4.0
openai>=1.0.0
psutil>=5.9.0
sqlparse>=0.4.4
aiohttp>=3.9.0
pydantic-settings
google-generativeai
pytest
pytest-asyncio
black
isort
flake8
fastapi
uvicorn
sqlmodel
tenacity
filelock
networkx
PyJWT
beautifulsoup4
PyYAML
websockets
cryptography
opentelemetry-api
opentelemetry-sdk
sentence_transformers
sqlalchemy
